backend/requirements.txt
# Backend requirements for the project
fastapi
uvicorn
pydantic
python-jose
passlib[bcrypt]
motor
